move decompositions to generic instructions, add circuit_to_instruction
