+++ qiskit/converters/__init__.py
@@ -15,3 +15,4 @@ from .circuits_to_qobj import circuits_to_qobj
 from .circuit_to_dag import circuit_to_dag
 from .dag_to_circuit import dag_to_circuit
 from .ast_to_dag import ast_to_dag
+from .circuit_to_instruction import circuit_to_instruction
